# web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "GitHubLastSyncResponse": {
        "description": "Last GitHub user-OAuth sync summary. Counts are honest; no secrets.",
        "properties": {
          "failed_count": {
            "title": "Failed Count",
            "type": "integer"
          },
          "new_count": {
            "title": "New Count",
            "type": "integer"
          },
          "removed_count": {
            "title": "Removed Count",
            "type": "integer"
          },
          "synced_at": {
            "title": "Synced At",
            "type": "string"
          },
          "unchanged_count": {
            "title": "Unchanged Count",
            "type": "integer"
          },
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "synced_at",
          "new_count",
          "updated_count",
          "unchanged_count",
          "removed_count",
          "failed_count"
        ],
        "title": "GitHubLastSyncResponse",
        "type": "object"
      },
      "GitHubProjectItemResponse": {
        "description": "One ProjectV2 row for the Hub picker.",
        "properties": {
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "owner_login": {
            "title": "Owner Login",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["owner_login", "number", "title"],
        "title": "GitHubProjectItemResponse",
        "type": "object"
      },
      "GitHubProjectPageResponse": {
        "description": "One page of ProjectV2 projects for the Hub picker.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GitHubProjectItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items"],
        "title": "GitHubProjectPageResponse",
        "type": "object"
      },
...
      "GitHubRepoPageResponse": {
        "description": "One page of repositories for the Hub picker.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GitHubRepoItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          }
        },
        "required": ["items"],
        "title": "GitHubRepoPageResponse",
        "type": "object"
      },
      "GitHubSelectionRequest": {
        "description": "Replace the saved GitHub sync selection.",
        "properties": {
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "project_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Number"
          },
          "project_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Owner"
          },
          "repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo"
          }
        },
        "title": "GitHubSelectionRequest",
        "type": "object"
      },
      "GitHubSelectionResponse": {
        "description": "Saved Hub sync targets (no tokens).",
        "properties": {
          "connector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Id"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "project_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Number"
          },
          "project_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Owner"
          },
          "repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo"
          }
        },
        "title": "GitHubSelectionResponse",
        "type": "object"
      },
      "GitHubStatusResponse": {
        "description": "GitHub connector presence and user OAuth connection (no secrets).",
        "properties": {
          "account_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Login"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "connected": {
            "default": false,
            "title": "Connected",
            "type": "boolean"
          },
          "connection_state": {
            "default": "disconnected",
            "title": "Connection State",
            "type": "string"
          },
          "document_count": {
            "default": 0,
            "title": "Document Count",
            "type": "integer"
          },
          "last_sync": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GitHubLastSyncResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "oauth_ready": {
            "default": false,
            "title": "Oauth Ready",
            "type": "boolean"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "project_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Number"
          },
          "project_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Owner"
          },
          "reauthorization_required": {
            "default": false,
            "title": "Reauthorization Required",
            "type": "boolean"
          },
          "repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo"
          },
          "setup_required": {
            "default": false,
            "title": "Setup Required",
            "type": "boolean"
          },
          "sync_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Scope"
          }
        },
        "required": ["configured", "available"],
        "title": "GitHubStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details plus Kernector extensions.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["type", "title", "status", "detail", "code"],
        "title": "Problem",
        "type": "object"
      },
      "ProblemError": {
        "description": "Field-level Problem Details extension entry.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          }
        },
        "required": ["pointer", "detail"],
        "title": "ProblemError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Minimal versioned HTTP adapter over composition. Unknown paths return RFC 9457 Problem Details with status 404 (app-level; not declared per operation).",
    "title": "Kernector API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/connectors/github": {
      "delete": {
        "description": "Revoke and delete the stored user grant. Synced documents are removed.",
        "operationId": "github_connector_disconnect_api_v1_connectors_github_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Github Connector Disconnect",
        "tags": ["connectors"]
      },
      "get": {
        "description": "Return presentation-safe GitHub PAT flags and user OAuth connection.",
        "operationId": "github_connector_status_api_v1_connectors_github_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Github Connector Status",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/last-sync": {
      "get": {
        "description": "Return the last persisted GitHub sync summary, if present.",
        "operationId": "github_connector_last_sync_api_v1_connectors_github_last_sync_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GitHubLastSyncResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Github Connector Last Sync Api V1 Connectors Github Last Sync Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Github Connector Last Sync",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/oauth/callback": {
      "get": {
        "description": "Validate state, exchange the code server-side, and return to the Hub.",
        "operationId": "github_oauth_callback_api_v1_connectors_github_oauth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Github Oauth Callback",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/oauth/start": {
      "get": {
        "description": "Issue CSRF state and redirect the browser to GitHub, or back to the Hub.",
        "operationId": "github_oauth_start_api_v1_connectors_github_oauth_start_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Github Oauth Start",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/projects": {
      "get": {
        "description": "List ProjectV2 projects for a login using the stored grant.",
        "operationId": "github_connector_projects_api_v1_connectors_github_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner_login",
            "required": true,
            "schema": {
              "title": "Owner Login",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubProjectPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Provider error"
          }
        },
        "summary": "Github Connector Projects",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/repos": {
      "get": {
        "description": "List repositories visible to the stored grant for the Hub picker.",
        "operationId": "github_connector_repos_api_v1_connectors_github_repos_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubRepoPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Provider error"
          }
        },
        "summary": "Github Connector Repos",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/selection": {
      "get": {
        "description": "Return the saved repository and optional ProjectV2 selection.",
        "operationId": "github_connector_get_selection_api_v1_connectors_github_selection_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubSelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Github Connector Get Selection",
        "tags": ["connectors"]
      },
      "put": {
        "description": "Validate access and atomically replace the saved GitHub selection.",
        "operationId": "github_connector_put_selection_api_v1_connectors_github_selection_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubSelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Provider error"
          }
        },
        "summary": "Github Connector Put Selection",
        "tags": ["connectors"]
      }
    },
    "/api/v1/connectors/github/sync": {
      "post": {
        "description": "Synchronize GitHub for the stored user OAuth grant.",
        "operationId": "github_connector_sync_api_v1_connectors_github_sync_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Server error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Provider error"
          }
        },
        "summary": "Github Connector Sync",
        "tags": ["connectors"]
      }
    },
...
  }
}